refactor: 카테고리 관련 기능들 auction패키지로 이동
